Clean up release generation, add prefix option to installer script.
* Remove unnecessary .framework and .framework.dSYM from the archive roots.
* Unify removal of these, so it's not in each of the targets.

+++ trunk/tools/install.sh
@@ -3,7 +3,9 @@
 # Installer script for ATVFiles @VERSION@ via Makeself.
 
 SRCDIR="$PWD"
-ATVFILES_DEST="/System/Library/CoreServices/Finder.app/Contents/PlugIns"
+COMMAND="${1:-install}"
+PREFIX="${2:-}"
+ATVFILES_DEST="${PREFIX}/System/Library/CoreServices/Finder.app/Contents/PlugIns"
 ATVFILES_NAME="ATVFiles.frappliance"
 ARCHIVE_NAME="@ARCHIVE_NAME@"
 
@@ -30,14 +32,25 @@ if mount | grep ' on / '  | grep -q 'read-only'; then
   /sbin/mount -uw /
 fi
 
-if [ "$1" = "uninstall" ]; then
+if [ "$COMMAND" = "uninstall" ]; then
   echo "== Removing ATVFiles"
   /bin/rm -rf "$ATVFILES_DEST/$ATVFILES_NAME" || die "Unable to uninstall ATVFiles"
 
   echo "ATVFiles successfully uninstalled."
   echo
   echo "Finder must be restarted in order to complete the installation."
-else
+elif [ "$COMMAND" = "help" ]; then
+  echo "Usage: $0 [action] [prefix]"
+  echo
+  echo "Install ATVFiles @VERSION@, optionally to a prefix"
+  echo
+  echo "Where action is:"
+  echo "  install       Install ATVFiles"
+  echo "  uninstall     Uninstall ATVFiles"
+  echo
+  echo "prefix is the root to a mounted install.  If specified, install will be automated"
+  echo "and will not restart Finder."
+elif [ "$COMMAND" = "install" ]; then
   
   # move atvfiles existing out of way
   if [ -d "$ATVFILES_DEST/$ATVFILES_NAME" ]; then
@@ -52,18 +65,20 @@ else
   
   echo "ATVFiles successfully installed."
   echo
-  echo "Finder must be restarted in order to complete the installation."
-fi
 
-# PRompt to restart finder
-echo
-echo -n "Would you like to do this now? (Y/n) "
-read -e restartfinder
-if [[ "$restartfinder" == "" || "$restartfinder" == "Y" || "$restartfinder" == "y" ]]; then
-  echo
-  echo "== Restarting Finder"
-  
-  kill `ps awx | grep [F]inder | awk '{print $1}'`
+  # Prompt to restart finder
+  if [ "$PREFIX" = "" ]; then
+    echo "Finder must be restarted in order to complete the installation."
+    echo
+    echo -n "Would you like to do this now? (Y/n) "
+    read -e restartfinder
+    if [[ "$restartfinder" == "" || "$restartfinder" == "Y" || "$restartfinder" == "y" ]]; then
+      echo
+      echo "== Restarting Finder"
+
+      kill `ps awx | grep [F]inder | awk '{print $1}'`
+    fi
+  fi # prefix empty
 fi
 
 # remount root as we found it
